+++ src/app/(views)/viewer/page.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, Suspense, useState } from "react";
 import { useSearchParams, useRouter } from "next/navigation";
 import type { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
 import SplatViewer from "./components/SplatViewer";
-import SceneItem from "../models/SceneItem";
+import SceneItem from "../../lib/definitions/SceneItem";
 
 const Viewer: React.FC = () => {
   const router = useRouter();