+++ src/app/api/fetchVideoItems/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { db } from "../../lib/db/db";
 import { splats } from "../../lib/db/schema";
-import VideoItem from "../../models/VideoItem";
+import VideoItem from "../../lib/definitions/VideoItem";
 
 export const dynamic = "force-dynamic";
 export const revalidate = 0;