+++ src/app/lib/definitions/SplatUploadPayload.ts
@@ -1,0 +1,6 @@
+export default interface SplatUploadPayload {
+    name: string,
+    description: string | null,
+    splatFile: File,
+    videoFile: File,
+}